+++ changelog.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Methods that have the `[DoesNotReturn]` attribute are now mocked such that they will throw `DoesNotReturnException` (this is to address the `CS8763` warning).
+- Methods that have the `[DoesNotReturn]` attribute are now mocked such that they will throw `DoesNotReturnException`. This is to address the `CS8763` warning (issue [#169](https://github.com/JasonBock/Rocks/issues/169)). 
+- If the mock type is an interface and it contains methods that match accessible methods from the `object` type (e.g. `Equals(T? other)` from `IEquatable<T>`), the methods will be implemented in the mock explicitly. This is to address the `CS0114` and `CS0108` warnings (issue [#169](https://github.com/JasonBock/Rocks/issues/169)).
 
 ## [6.4.4] - 2022-06-27